# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 6321747fc97c70218d65cc7e
  title: SchoolManagement
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/teacher/{id}":
    delete:
      x-skaffolder-id: 632176bfc97c70218d65cd3c
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 632175d0c97c70218d65cd14
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 632176bec97c70218d65cd37
      x-skaffolder-name: get
      x-skaffolder-id-resource: 632175d0c97c70218d65cd14
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/teacher"
    post:
      x-skaffolder-id: 632176bec97c70218d65cd35
      x-skaffolder-name: update
      x-skaffolder-id-resource: 632175d0c97c70218d65cd14
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/teacher"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/teacher"
  "/teacher/findBy_courses/{key}":
    get:
      x-skaffolder-id: 632176c5c97c70218d65cd47
      x-skaffolder-name: findBy_courses
      x-skaffolder-id-resource: 632175d0c97c70218d65cd14
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: findBy_courses
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_courses
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_courses/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_courses
      tags:
        - teacher
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/teacher"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    teacher:
      x-skaffolder-id: 632175d0c97c70218d65cd14
      x-skaffolder-id-db: 6321747fc97c70218d65cc7f
      x-skaffolder-id-entity: 632175d0c97c70218d65cd11
      x-skaffolder-url: /teacher
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        fname:
          type: string
          x-skaffolder-id-attr: 632175d0c97c70218d65cd12
          x-skaffolder-type: String
        lname:
          type: string
          x-skaffolder-id-attr: 632175d0c97c70218d65cd13
          x-skaffolder-type: String
      x-skaffolder-relations:
        _courses:
          x-skaffolder-id: 6321764cc97c70218d65cd24
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 632175d0c97c70218d65cd11
          x-skaffolder-ent2: 632175eac97c70218d65cd17
      required:
        - _id
...